# decoder: parsing optimisations