Optimisations of the parsing: avoid calling functions of the `bytes` package for operations on just one or two bytes.