@alevinval 2016-10-18
I cannot believe it, another contributor !
I am very sorry I just pushed something and now I was checking it and saw your pull request. It looks like I caused some conflicts !