@alevinval 2016-10-18
I can try to solve them and make a PR to your fork.