@dolmen 2016-10-18
Don't worry about the conflicts: I resolved them and rebased. Check my updated branch.