@alevinval 2016-10-18
I did it via CLI at the end. Bad timing all around.
I only added a few comments where you replaced standard library by the optimised code.

Thank you !