@alevinval 2016-10-18
FYI
avoiding an extra String() call I noticed + your commits afterwards:

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkDecodeEmptyEvent-4                    1399          804           -42.53%
BenchmarkDecodeEmptyEventWithIgnoredLine-4     1519          892           -41.28%
BenchmarkDecodeShortEvent-4                    1437          898           -37.51%
BenchmarkDecode8kEvent-4                       8983          6199          -30.99%
BenchmarkDecode16kEvent-4                      14820         10309         -30.44%
```