@dolmen 2016-10-22
>  cannot believe it, another contributor !

Actually this is not a coincidence: we are working in the same office.